BHBC-2224: Fix integer cast zod schema

+++ api/src/repositories/artifact-repository.ts
@@ -47,7 +47,7 @@ export class ArtifactRepository extends BaseRepository {
 
     const sqlStatement = SQL`
       SELECT
-        NEXTVAL('artifact_seq') AS artifact_id
+        NEXTVAL('artifact_seq')::integer AS artifact_id
       FROM
         GENERATE_SERIES(1, ${count});
     `;
